playerimpl class was moved to repisitory package
